Remove InputLoopService from the Core_9 sample
