+++ samples/saga/simple/Core_9/Sample/Program.cs
@@ -1,46 +1,53 @@
 using System;
-using System.Threading.Tasks;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 using NServiceBus;
-using Sample;
 
-class Program
+var builder = Host.CreateApplicationBuilder(args);
+
+#region config
+
+var endpointConfiguration = new EndpointConfiguration("Samples.SimpleSaga");
+endpointConfiguration.UsePersistence<LearningPersistence>();
+endpointConfiguration.UseTransport(new LearningTransport());
+endpointConfiguration.UseSerialization<SystemJsonSerializer>();
+
+#endregion
+
+builder.UseNServiceBus(endpointConfiguration);
+
+Console.Title = "Server";
+
+var host = builder.Build();
+
+await host.StartAsync();
+
+var messageSession = host.Services.GetRequiredService<IMessageSession>();
+
+Console.WriteLine();
+Console.WriteLine("Storage locations:");
+Console.WriteLine($"Learning Persister: {LearningLocationHelper.SagaDirectory}");
+Console.WriteLine($"Learning Transport: {LearningLocationHelper.TransportDirectory}");
+
+Console.WriteLine();
+Console.WriteLine("Press 'Enter' to send a StartOrder message");
+Console.WriteLine();
+
+
+while (true)
 {
-    public static async Task Main(string[] args)
+    if (Console.ReadKey().Key != ConsoleKey.Enter)
     {
-        await CreateHostBuilder(args).Build().RunAsync();
+        break;
     }
+    var orderId = Guid.NewGuid();
+    var startOrder = new StartOrder
+    {
+        OrderId = orderId
+    };
+    await messageSession.SendLocal(startOrder);
 
-    public static IHostBuilder CreateHostBuilder(string[] args) =>
-     Host.CreateDefaultBuilder(args)
-         .ConfigureServices((hostContext, services) =>
-         {
-             Console.Title = "Server";
-             services.AddHostedService<InputLoopService>();
-
-         }).UseNServiceBus(x =>
-         {
-             Console.Title = "SimpleSaga";
-             var endpointConfiguration = new EndpointConfiguration("Samples.SimpleSaga");
-
-             #region config
-
-             endpointConfiguration.UsePersistence<LearningPersistence>();
-             endpointConfiguration.UseTransport(new LearningTransport());
-             endpointConfiguration.UseSerialization<SystemJsonSerializer>();
-
-             #endregion
-
-
-             Console.WriteLine();
-             Console.WriteLine("Storage locations:");
-             Console.WriteLine($"Learning Persister: {LearningLocationHelper.SagaDirectory}");
-             Console.WriteLine($"Learning Transport: {LearningLocationHelper.TransportDirectory}");
-
-             Console.WriteLine();
-             Console.WriteLine("Press 'Enter' to send a StartOrder message");
-             return endpointConfiguration;
-         });
-
+    Console.WriteLine($"Sent StartOrder with OrderId {orderId}.");
 }
+
+await host.StopAsync();
